+++ RayTracedGGX/XUSG/Core/XUSGPipelineLayout.h
@@ -43,7 +43,7 @@ namespace XUSG
 
 			void SetShaderStage(uint32_t index, Shader::Stage stage);
 			void SetRange(uint32_t index, DescriptorType type, uint32_t num, uint32_t baseBinding,
-				uint32_t space = 0, uint8_t flags = 0);
+				uint32_t space = 0, uint32_t flags = 0);
 			void SetConstants(uint32_t index, uint32_t num32BitValues, uint32_t binding,
 				uint32_t space = 0, Shader::Stage stage = Shader::Stage::ALL);
 			void SetRootSRV(uint32_t index, uint32_t binding,
